bi-directional s3 syncing tool
Implements bi-directional syncing of a directory
to an S3 bucket, given a manifest file detailing
which files we care about syncing.

We short-circuit if the file hashes are equal,
and only sync up if the local file is newer. Similarly,
we only sync down if the hashes are different and the remote
file is newer.

We always check the remote first.